Renaming component files and refactor prettier config into its own file
