+++ frontend/src/components/content-component/content-component.scss
