+++ frontend/src/components/query-component/query-component.scss
